chore(clang-tidy): P1 burndown #4 — eliminate bugprone-narrowing-conversions
Drops `bugprone-narrowing-conversions` from 19 → 0. All real fixes,
no NOLINTs — every site became an explicit `static_cast` to the
exact target type. The casts are behaviourally identical to the
implicit conversions they replace: in every case the source values
are bounded by Qt container sizes or screen geometry that fit in an
`int` or a `float` by construction.

apps/etx Qt UI code (18 sites)

  Qt 6 returns `qsizetype` (signed long long) from container size()
  / indexOf(). Several APIs (setRowCount, QPolyline::drawPolyline,
  history index tracking) want `int`; we cast at the boundary.

  Casts to `int`:
    apps/etx/src/screens/help.cpp:73
    apps/etx/src/screens/metrics.cpp:134
    apps/etx/src/theme.cpp:28-29        (indexOf + size)
    apps/etx/src/widgets/bento_layout.cpp:38
    apps/etx/src/widgets/breadcrumb.cpp:40
    apps/etx/src/widgets/chat_input.cpp:177
    apps/etx/src/widgets/chat_input.cpp:302  (qBound 3rd arg)
    apps/etx/src/widgets/sparkline.cpp:101   (drawPolyline count)
    apps/etx/src/widgets/tip_bar.cpp:62      (modulo size)

  Casts to `double` (denominators in pixel-step math):
    apps/etx/src/widgets/header.cpp:53
    apps/etx/src/widgets/resource_monitor.cpp:41
    apps/etx/src/widgets/sparkline.cpp:73

  Casts to `float` (mixed int/float pixel math):
    apps/etx/src/widgets/resource_monitor.cpp:46
       (values_[i] → float, height() → float, both in one expr)
    apps/etx/src/widgets/scroll_minimap.cpp:47-48
       (viewport_start_ / viewport_size_ * h, h → float)
    apps/etx/src/widgets/scroll_minimap.cpp:55
       (clamped * h, h → float)

Runtime test (1 site)

  libs/runtime/tests/test_session_store.cpp:62 — `f.write(...,
  packed.size())` converted `size_type` (unsigned long) to
  `streamsize` (long). Wrapped in `static_cast<std::streamsize>`.

Verification
  - cmake --build cmake-build-san --target euxis-etx-tests
    euxis-runtime-cpp_tests  clean
  - ctest -R "(etx|runtime)"   2/2 pass
  - clang-tidy 22.1.7 on every touched file:
    `bugprone-narrowing-conversions` count is 0.

Continues #42. Stacks cleanly on top of #110, #111, #112 — no
file overlap.
---
THE ARCHITECT ᛫ Sebastien Rousseau ᛫ https://sebastienrousseau.com
THE ENGINE ᛞ EUXIS ᛫ Enterprise Unified Execution Intelligence System ᛫ https://euxis.co

+++ apps/etx/src/widgets/resource_monitor.cpp
@@ -38,12 +38,12 @@ protected:
         if (values_.size() < 2) return;
 
         float max_val = 100.0f;
-        double step_x = static_cast<double>(width()) / (values_.size() - 1);
+        double step_x = static_cast<double>(width()) / static_cast<double>(values_.size() - 1);
 
         QVector<QPointF> points;
         for (int i = 0; i < values_.size(); ++i) {
             double x = i * step_x;
-            double y = height() - (values_[i] / max_val * height());
+            double y = height() - (static_cast<float>(values_[i]) / max_val * static_cast<float>(height()));
             points.append(QPointF(x, y));
         }
